UI Improvements Move authentication logic into custom hook Add branding log, favicon and change the name of the application
