+++ src/components/AppNoAuth/AppNoAuth.tsx
@@ -1,20 +1,14 @@
 import React from 'react';
-import ChallengeList from '../ChallengeList';
-import ChallengeProfile from '../ChallengeProfile';
-import EcoverseProfile from '../EcoverseProfile';
+import EcoverseContainer from '../../containers/EcoverseContainer';
+import Header from '../Header';
 import './AppNoAuth.css';
 
 const AppNoAuth = (): React.ReactElement => {
-  const [id, setId] = React.useState(2);
-  const handleIdChange = React.useCallback(newId => {
-    setId(newId);
-  }, []);
-
+  const noOp = () => {};
   return (
     <div className="AppNoAuth">
-      <EcoverseProfile />
-      <ChallengeList handleIdChange={handleIdChange} />
-      <ChallengeProfile id={id} />
+      <Header userName="No Logged User" isAuthenticated={false} onSignIn={noOp} onSignOut={noOp} />
+      <EcoverseContainer />
     </div>
   );
 };
