+++ src/components/Header/Header.css
@@ -1,0 +1,3 @@
+.logo-svg {
+  height: 30px;
+}
