+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useClientConfig';
+export * from './useAuthentication';
